+++ url2parquet/f_process_and_save_parquet.py
@@ -9,12 +9,11 @@ from datetime import datetime, timedelta
 from tqdm import tqdm
 from glob import glob
 from io import BytesIO
-from nc_read_downloaded import nc_reader
 from changable_things import data_types, start_date, end_date, parquet_directory, nc_file_path, merged_directory, size_ranges, download_period, total_steps
 import stat
 
 
-# 이미지 데이터를 Parquet로 저장하는 함수
+# function to process and save parquet
 def process_and_save_parquet(data_list, data_types, parquet_save_path, size_ranges):
     if parquet_directory:
         if not os.path.exists(parquet_directory):
@@ -42,19 +41,19 @@ def process_and_save_parquet(data_list, data_types, parquet_save_path, size_rang
             print(f"{rows}x{cols} 크기의 데이터는 범위 정의가 없습니다. 스킵합니다.")
             continue
 
-        # 필요한 범위 자르기
+        # cut data by range(what we really need)
         all_data_for_type = [data[y_range[0]:y_range[1] +1, x_range[0]:x_range[1] +1] for data in all_data_for_type]
 
-        # 평균 계산
+        # calculate average
         averaged_data = np.nanmean(np.stack(all_data_for_type), axis=0)
 
-        # 소수점 자리수 제한
+        # round to 3 decimal places
         averaged_data = np.round(averaged_data, decimals=3)
 
-        # float16으로 변환
+        # convert to float16
         averaged_data = averaged_data.astype(np.float16)
 
-        # 좌표 생성 및 결과 데이터 저장
+        # data raveling(stretching)?
         rows, cols = averaged_data.shape  #cut data의 행의 개수, 열의 개수 -> 행의 개수면 y좌표수랑 같겠죠..
         x, y = np.meshgrid(range(cols), range(rows))
         flattened_data = pd.DataFrame({
@@ -69,7 +68,7 @@ def process_and_save_parquet(data_list, data_types, parquet_save_path, size_rang
         else:
             size_grouped_data[size_cut] = pd.merge(size_grouped_data[size_cut], flattened_data, on=['x', 'y'], how='outer')
 
-    # Parquet 저장
+    # save hourly parquet files
     for size_cut, combined_data in size_grouped_data.items():
         parquet_save_path_with_size = f"{parquet_save_path}_{size_cut}.parquet"
         combined_data.to_parquet(parquet_save_path_with_size, engine='pyarrow', compression='snappy')
